requirements.txt
streamlit>=1.37
faiss-cpu>=1.7.4
sentence-transformers>=3.0.0
pydantic>=2.0.0
pyyaml>=6.0.1
python-dotenv>=1.0.1
matplotlib>=3.8.0
requests>=2.31.0
